+++ Runtime/Extensions/StringExtensions.cs
@@ -1,10 +1,16 @@
 ﻿using UnityEngine;
 
-namespace DapperDino.DapperTools
+namespace DapperDino.DapperTools.Extensions
 {
     public static class StringExtensions
     {
-        public static string AsColour(this string text, Color colour)
+        /// <summary>
+        /// Return the text ready to be displayed on the UI as the supplied colour
+        /// </summary>
+        /// <param name="text"></param>
+        /// <param name="colour"></param>
+        /// <returns></returns>
+        public static string WithColour(this string text, Color colour)
         {
             return $"<color=#{ColorUtility.ToHtmlStringRGB(colour)}>{text}</color>";
         }